change twitter_user_id data type on TrollCandidate

+++ db/migrate/20170104045550_add_twitter_user_id_to_troll_candidate.rb
@@ -1,6 +1,6 @@
 class AddTwitterUserIdToTrollCandidate < ActiveRecord::Migration[5.0]
   def change
-    add_column :troll_candidates, :twitter_user_id, :integer
+    add_column :troll_candidates, :twitter_user_id, :bigint
 
     add_index :troll_candidates, :twitter_user_id
   end
